task-runner: Rename frostfire/settings.ts to setting.ts
