+++ src/scene/scenes/frostfire/setting.ts
